Added verification/flagging of bathrooms and Omniauth authentication to allow Facebook/Google logging in
Added the necessary columns and functions to allow users to verify or
flag bathrooms. Current logic will verify the bathroom at a total of 5
positive verifications, and will delete the bathroom with a total of -5
verifications.
I also added the Omniauth Gem in order to be able to allow users to
login to the app with their Google or Facebook accounts, however I am
not able to get it working locally yet, so it needs debugging. This
might be due to errors in inputting data into Google and Facebook when
getting the App Id and Secret Keys…